+++ src/city/world/__init__.py
@@ -2,6 +2,5 @@ from . import (
 	assets,
     data,
     maker,
-    terrain_painter,
     tile_set
 )
